add ability to access git from interpreter, give skeleton user interface class
